# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HypeSpace Event Management SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members of an organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: inviteTeamMember
      tags: [team]
      summary: Invite a team member
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteTeamMemberBody"
      responses:
        "201":
          description: Invited team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /organizations/{orgId}/team/{memberId}:
    put:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member role
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberBody"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: removeTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /organizations/{orgId}/social-posts:
    get:
      operationId: listSocialPosts
      tags: [social]
      summary: List social media posts for organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of social posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPost"
    post:
      operationId: createSocialPost
      tags: [social]
      summary: Create a social media post
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSocialPostBody"
      responses:
        "201":
          description: Created social post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
  /organizations/{orgId}/social-posts/{postId}:
    put:
      operationId: updateSocialPost
      tags: [social]
      summary: Update a social media post
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSocialPostBody"
      responses:
        "200":
          description: Updated social post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
    delete:
      operationId: deleteSocialPost
      tags: [social]
      summary: Delete a social media post
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /organizations/{orgId}/social-posts/{postId}/publish:
    post:
      operationId: publishSocialPost
      tags: [social]
      summary: Publish a social media post
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
...
  /organizations/{orgId}/events/{eventId}/reminders:
    get:
      operationId: listReminders
      tags: [campaigns]
      summary: List reminders for an event
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [campaigns]
      summary: Create a reminder for an event
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderBody"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
components:
  schemas:
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        organizationId:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, admin, manager, member]
        status:
          type: string
          enum: [active, invited, suspended]
        joinedAt:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required:
        - id
        - userId
        - organizationId
        - email
        - name
        - role
        - status
    InviteTeamMemberBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, member]
      required:
        - email
        - name
        - role
    UpdateTeamMemberBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, manager, member]
        status:
          type: string
          enum: [active, suspended]
...
    SocialPost:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        eventId:
          type: integer
          nullable: true
        platform:
          type: string
          enum: [twitter, linkedin, facebook, instagram]
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, scheduled, published, failed]
        scheduledAt:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - organizationId
        - platform
        - content
        - status
        - createdAt
    CreateSocialPostBody:
      type: object
      properties:
        eventId:
          type: integer
          nullable: true
        platform:
          type: string
          enum: [twitter, linkedin, facebook, instagram]
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        scheduledAt:
          type: string
          nullable: true
      required:
        - platform
        - content
    UpdateSocialPostBody:
      type: object
      properties:
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        scheduledAt:
          type: string
          nullable: true
    Reminder:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        type:
          type: string
          enum: [before_event, after_event, custom]
        offsetHours:
          type: integer
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, sent, failed]
        scheduledAt:
          type: string
          nullable: true
        sentAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - eventId
        - type
        - offsetHours
        - subject
        - message
        - status
        - createdAt
    CreateReminderBody:
      type: object
      properties:
        type:
          type: string
          enum: [before_event, after_event, custom]
        offsetHours:
          type: integer
        subject:
          type: string
        message:
          type: string
      required:
        - type
        - offsetHours
        - subject
        - message
...