# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HypeSpace Event Management SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderStatus } from "./reminderStatus";
import type { ReminderType } from "./reminderType";

export interface Reminder {
  id: number;
  eventId: number;
  type: ReminderType;
  offsetHours: number;
  subject: string;
  message: string;
  status: ReminderStatus;
  scheduledAt?: string | null;
  sentAt?: string | null;
  createdAt: string;
}